converted/tier1-frameworks/fastapi/requirements.txt
# Python dependencies for FastAPI on Elide

# Core FastAPI functionality (for reference and polyglot imports)
fastapi>=0.103.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Async support
asyncio>=3.4.3

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning (optional, for ML examples)
scikit-learn>=1.3.0
scipy>=1.11.0

# NLP (optional, for NLP examples)
spacy>=3.6.0
textblob>=0.17.1
nltk>=3.8.0

# Database (optional, for database examples)
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0

# Date/time utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
